% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineups_searcher.R
\name{lineups_searcher}
\alias{lineups_searcher}
\title{Statistics searcher}
\usage{
lineups_searcher(df1, n, p1, p2, p3, p4)
}
\arguments{
\item{df1}{Should be a Data Frame. This parameter has to be in the format provided by the lineups_advance_stats() function.}

\item{n}{Should be a numer. It represents the number of player to be found.}

\item{p1}{Should be a String. Represents the name of the first player to be found.}

\item{p2}{Should be a String. Represents the name of the second player to be found.}

\item{p3}{Should be a String. Represents the name of a player to be found.}

\item{p4}{Should be a String. Represents the name of a player to be found.}
}
\value{
Data frame with the statistics of the lineups where the entered players appear
}
\description{
The function allows the statistical search of the lineups where the entered players appear.
}
\details{
\itemize{
                \item The function allows you to search for paint players both in basic statistics and in extended statistics.
                \item The values allowed by n are 1, 2, 3 and 4.The number entered in N must be equal to the number of players searched.
                \item The name entered in the function must be the same as the one inside the data frame.
        }
}
\examples{

df1 <- data.frame("PG" = c("James","Rondo"),"SG" = c("Green","Caruso"),
"SF" = c("Caldwell","Kuzma"), "PF" = c("Davis","Davis"),
"C" = c("Howard ","Howard"),"MP" = c(7,1), "FG " = c(4,0),
"FGA " = c(7,0),"Percentage FG" = c(0.571,0),
"X3P  " = c(0,0),"X3PA  " = c(2,0),"Percentage 3P" = c(0,0),
"X2P " = c(4,0), "X2PA " = c(5,0), "Percentage 2P" = c(0.8,0),
"FT " = c(1,0), "FTA " = c(3,0), "Percentage FT" = c(0.333,0),
"ORB " = c(2,0), "DRB " = c(5,0),"TRB " = c(7,0), "AST " = c(2,0),
"STL " = c(1,0), "BLK " = c(0,0),"TOV " = c(7,2), "PF" = c(1,0),
"PLUS" = c(9,0),"MINUS" = c(17,3),"P/M" = c(-8,-3))

n <- 2

p1 <- "James"

p2 <- "Davis"

p3 <- ""

p4 <- ""


lineups_searcher(df1,n,p1,p2,p3,p4)

}
\author{
Fco Javier Cantero \email{fco.cantero@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
