% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_games_adder.R
\name{play_games_adder}
\alias{play_games_adder}
\title{Play games adder}
\usage{
play_games_adder(df1, df2)
}
\arguments{
\item{df1}{Should be a Data Frame that represents the first set of play's statistics. It has to be in the format provided by the play_data_adjustment() function.}

\item{df2}{Should be a Data Frame that represents the second set of play's statistics. It has to be in the format provided by the play_data_adjustment() function.}
}
\value{
Data frame with the sum of the statistics of the other two entered data frame.
}
\description{
The function allows to perform the sums of two data.frames with the same format adopted after being transformed by play_data_adjustment() function,
}
\details{
The function will work correctly when the name of the players is the same, in case it is different it will take the players as different.
}
\examples{

df1 <- data.frame("Name" = c("Sabonis ","Team"), "GP" = c(62,71),
"PTS" = c(387,0), "FG" = c(155,1), "FGA" = c(281,1),
"FGA Percentage" = c(0.552,1),"3P" = c(6,1),"3PA" = c(18,1),
"3P Percentage" = c(0.333,1),"2P" = c(149,0),"2PA" = c(263,0),
"2P Percentage" = c(0.567,0),"FT" = c(39,1),  "FTA" = c(53,1),
"FT Percentage" = c(0.736,1),  "ANDONE" = c(12,1), "AST" = c(0,1),
"TOV" = c(27,1))

df2 <- data.frame("Name" = c("Sabonis ","Team"), "GP" = c(62,71),
"PTS" = c(387,0), "FG" = c(155,1), "FGA" = c(281,1),
"FGA Percentage" = c(0.552,1),"3P" = c(6,1),"3PA" = c(18,1),
"3P Percentage" = c(0.333,1),"2P" = c(149,0),"2PA" = c(263,0),
"2P Percentage" = c(0.567,0),"FT" = c(39,1),  "FTA" = c(53,1),
"FT Percentage" = c(0.736,1),  "ANDONE" = c(12,1), "AST" = c(0,1),
"TOV" = c(27,1))

play_games_adder(df1,df2)


}
