% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_stats_per_game.R
\name{play_stats_per_game}
\alias{play_stats_per_game}
\title{Play stats per game}
\usage{
play_stats_per_game(df1)
}
\arguments{
\item{df1}{Should be a Data Frame that represents the play's statistics. The parameter has to be in the format provided by the play_data_adjustment() function.}
}
\value{
Data frame with play statistics per game
}
\description{
The function allows the calculation of play statistics per game.
}
\details{
The calculation is made with the number of games played by the player.
}
\examples{

df1 <- data.frame("Name" = c("Sabonis ","Team"), "GP" = c(62,71),
"PTS" = c(387,0), "FG" = c(155,1), "FGA" = c(281,1),
"FGA Percentage" = c(0.552,1),"3P" = c(6,1),"3PA" = c(18,1),
"3P Percentage" = c(0.333,1),"2P" = c(149,0),"2PA" = c(263,0),
"2P Percentage" = c(0.567,0),"FT" = c(39,1),  "FTA" = c(53,1),
"FT Percentage" = c(0.736,1),  "ANDONE" = c(12,1), "AST" = c(0,1),
"TOV" = c(27,1))

play_stats_per_game(df1)

}
\author{
Fco Javier Cantero \email{fco.cantero@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
