% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_stats_per_game.R
\name{team_stats_per_game}
\alias{team_stats_per_game}
\title{Team stats per game}
\usage{
team_stats_per_game(df1)
}
\arguments{
\item{df1}{Should be a Data Frame that represents the team statistics. This parameter has to be in the format provided by the team_stats() function.}
}
\value{
Data frame whit the team's statistics per game
}
\description{
The function allows the calculation of team's statistics per game.
}
\details{
The calculation is made with the number of games played by the team.
}
\examples{

df1 <- data.frame("G" = c(71), "MP" = c(17090), "FG" = c(3006),
"FGA" = c(6269),"Percentage FG" = c(0.48), "3P" = c(782),
"3PA" = c(2242), "Percentage 3P" = c(0.349),  "2P" = c(2224),
"2PA" = c(4027), "Percentage 2P" = c(0.552), "FT" = c(1260),
"FTA FG" = c(1728),  "Percentage FT" = c(0.729), "ORB" = c(757),
"DRB" = c(2490),"TRB" = c(3247),  "AST" = c(1803),  "STL" = c(612),
"BLK" = c(468),   "TOV" = c(1077),  "PF" = c(1471),  "PTS" = c(8054),
"+/-" = c(0))

team_stats_per_game(df1)

}
\author{
Fco Javier Cantero \email{fco.cantero@edu.uah.es}

Juan José Cuadrado \email{jjcg@uah.es}

Universidad de Alcalá de Henares
}
